application/app/requirements.txt
dateutils
flask
flask_httpauth 
requests
uwsgi
wheel
